fix(script): normalise uid comparisons via int32 in p_finduid + LookupPlayerByUID
The PushInt int32-cast fix (b56c207c) corrected combat varn round-trip
but introduced a follow-on regression: high-bit-set UIDs (~50% of
usernames) could no longer satisfy `if (p_finduid(uid) = true)` gates,
so every protected if_button branch fell through to its resync-only
fallback. Player-controls toggles (walk/run, auto-retaliate, etc.)
snapped back to their server-side state immediately after click —
mes("Someone else is fighting that") was the analogous combat symptom
before b56c207c closed it.

Cause: PushInt now normalises uids to their signed-int32 representation
on the script stack, but `ActivePlayer.UID()` and `Player.uid` still
return / store the composeUID-derived positive-Go-int form (matches TS
World.ts:937's `>>> 0` semantics). The direct equality comparisons in
handlePFindUID's self-reacquire fast-paths and in
Server.LookupPlayerByUID therefore compared 2147483649 against
-2147483647 → mismatch on every call.

TS sidesteps this two ways: (a) JS bitwise-coercion semantics
automatically toInt32 both sides of `===`/`==` so any uint32-shaped
JS-number compares equal to its int32-cast form, and (b)
getPlayerByUid (World.ts:1659-1673) decomposes uid into (slot, hash)
for a representation-independent lookup. Go has neither, so the fix
normalises at the comparison boundary: `int32(p.uid) == int32(uid)`
shares the bottom-32-bits + sign-extension contract with both
representations.

Sites touched:
- handlePFindUID self-reacquire fast-paths (two — slots 0 and 1)
- Server.LookupPlayerByUID
- mockPlayerLookup.LookupPlayerByUID (test fixture, in lock-step with
  production so tests using positive-uint32 keys keep working)

worldVarsView.LookupPlayerByUID delegates to Server's method, so it
inherits the fix transparently.

Two new tests pin the user-visible scenario:
- TestPFindUID_HighBitUid_FastPathReacquire: protected if_button's
  self-reacquire path. PlayerLookup nil so the fast-path must succeed
  standalone (would push 0 pre-fix).
- TestPFindUID_HighBitUid_LookupPath: non-protected path with target
  in mockPlayerLookup keyed by positive composeUID form.

Fifth chained "Go diverges from TS in a way comments never flagged"
this session — 30305239, 35657661, 6c7f7738, b56c207c, now this. The
PushInt fix's own ripple effect makes a tight closure pair with the
combat-bug fix it spawned from.

+++ pkg/script/handlers_player.go
@@ -1178,13 +1178,18 @@ func handlePFindUID(s *ScriptState) error {
 	uid := s.PopInt()
 
 	// Self-reacquire fast-path: already protected on this slot's player.
+	// Compare via int32-cast so the script-side stack representation
+	// (int32-normalised by PushInt per TS toInt32 parity, Numbers.ts:7)
+	// matches against ActivePlayer.UID() which returns the
+	// composeUID-derived uint32-bit-pattern as a positive Go int. Both
+	// sides agree on the bottom 32 bits; sign-extension reconciles them.
 	if operand == 0 {
-		if s.Pointers&PtrProtectedActivePlayer != 0 && s.Self != nil && s.Self.UID() == uid {
+		if s.Pointers&PtrProtectedActivePlayer != 0 && s.Self != nil && int32(s.Self.UID()) == int32(uid) {
 			s.PushInt(1)
 			return nil
 		}
 	} else {
-		if s.Pointers&PtrProtectedActivePlayer2 != 0 && s.Self2 != nil && s.Self2.UID() == uid {
+		if s.Pointers&PtrProtectedActivePlayer2 != 0 && s.Self2 != nil && int32(s.Self2.UID()) == int32(uid) {
 			s.PushInt(1)
 			return nil
 		}
